requirements.txt
PyYAML==6.0.1
boto3==1.34.0
kafka-python==2.0.2
click==8.1.7
colorama==0.4.6
